+++ src/data/v2/last_quote.rs
@@ -1,4 +1,4 @@
-// Copyright (C) 2021 The apca Developers
+// Copyright (C) 2021-2022 The apca Developers
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 use chrono::DateTime;
@@ -46,7 +46,7 @@ EndpointNoParse! {
   ],
   Err => GetError, [
     /// The provided symbol was invalid or not found.
-    /* 422 */ UNPROCESSABLE_ENTITY => InvalidSymbol,
+    /* 422 */ UNPROCESSABLE_ENTITY => InvalidInput,
   ]
 
   fn base_url() -> Option<Str> {
@@ -151,7 +151,7 @@ mod tests {
       .await
       .unwrap_err();
     match err {
-      RequestError::Endpoint(GetError::InvalidSymbol(_)) => (),
+      RequestError::Endpoint(GetError::InvalidInput(_)) => (),
       _ => panic!("Received unexpected error: {:?}", err),
     };
   }