+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Unreleased
 - Renamed various 422 HTTP status error variants to `InvalidInput`
 - Renamed `data::v2::bars::BarReq` to `BarsReq`
 - Changed various limit types to `usize`
+- Changed `api::v2::orders::OrdersReq::limit` to be an `Option`
 
 
 0.22.4