+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Unreleased
 ----------
 - Renamed `data::v2::bars::BarReq` to `BarsReq`
   - Deprecated `data::v2::bars::BarReq`
+- Introduced `data::v2::bars::BarsReqInit` type
 
 
 0.22.4